Refactor to separate package
