+++ ksqldb-engine/src/main/java/io/confluent/ksql/physical/scalablepush/PushPhysicalPlanBuilder.java
@@ -50,20 +50,17 @@ public class PushPhysicalPlanBuilder {
 
   private final ProcessingLogContext processingLogContext;
   private final PersistentQueryMetadata persistentQueryMetadata;
-  private final boolean expectingStartOfRegistryData;
   private final Stacker contextStacker;
   private final QueryId queryId;
 
   public PushPhysicalPlanBuilder(
       final ProcessingLogContext processingLogContext,
-      final PersistentQueryMetadata persistentQueryMetadata,
-      final boolean expectingStartOfRegistryData
+      final PersistentQueryMetadata persistentQueryMetadata
   ) {
     this.processingLogContext = Objects.requireNonNull(
         processingLogContext, "processingLogContext");
     this.persistentQueryMetadata = Objects.requireNonNull(
         persistentQueryMetadata, "persistentQueryMetadata");
-    this.expectingStartOfRegistryData = expectingStartOfRegistryData;
     this.contextStacker = new Stacker();
     queryId = uniqueQueryId();
   }
@@ -166,8 +163,7 @@ public class PushPhysicalPlanBuilder {
     final ScalablePushRegistry scalablePushRegistry =
         persistentQueryMetadata.getScalablePushRegistry()
         .orElseThrow(() -> new IllegalStateException("Scalable push registry cannot be found"));
-    return new PeekStreamOperator(scalablePushRegistry, logicalNode, queryId,
-        expectingStartOfRegistryData, token);
+    return new PeekStreamOperator(scalablePushRegistry, logicalNode, queryId, token);
   }
 
   private QueryId uniqueQueryId() {
