+++ ksqldb-engine/src/main/java/io/confluent/ksql/physical/scalablepush/PushRoutingOptions.java
@@ -25,9 +25,5 @@ public interface PushRoutingOptions {
   // If we should avoid skipping forwarding the request because it's already been forwarded.
   boolean getHasBeenForwarded();
 
-  // When a rebalance occurs and we connect to a new node, we don't want to miss anything, so we
-  // set this flag indicating we should error if this expectation isn't met.
-  boolean getExpectingStartOfRegistryData();
-
   Optional<String> getToken();
 }
