+++ ksqldb-engine/src/main/java/io/confluent/ksql/physical/scalablepush/consumer/ConsumerMetadata.java
@@ -1,0 +1,18 @@
+package io.confluent.ksql.physical.scalablepush.consumer;
+
+public class ConsumerMetadata implements AutoCloseable {
+
+  private final int numPartitions;
+
+  public ConsumerMetadata(final int numPartitions) {
+    this.numPartitions = numPartitions;
+  }
+
+  public int getNumPartitions() {
+    return numPartitions;
+  }
+
+  @Override
+  public void close() {
+  }
+}
